cpuRaw and ramRaw
Do not classify runtimes in ranges of available resources.
All runtimes are put in the cpuRaw/1 and ramRaw/1000.
When requesting nodes for some actor, we return all available and filter
them after.

Also, CpuTotal accepts now any value. But when using cpuTotal in
deployment files, it must be in accept range of values.

Minor fixes in tests