New cooldown implementation.
Set the parameter by runtime to work in the centralized version.
As the app may contain actors from many runtimes, consider the runtime
that asked for migration as origin for the application.

In practice, now it will be the sink actor.
In the end, it is the same behavior we had before...